requirements.txt
ultralytics
icrawler
python-dotenv
pyyaml
opencv-python
